Fixed Create Table Error
-Table for Words DB won't be created if it already exists
-Fixed minor bug where only 4 letter words were being populated into the DB. Now only 5 letter words are being entered as per project reqs
-Separated DB Initialization into another Python file
